Implement SabHeroEstimator class with methods for retrieving rates, multipliers, house styles, and settings; update service provider to register SabHeroEstimator singleton; remove unused filters placeholder in SettingResource table configuration.
