+++ src/Filament/Resources/SettingResource.php
@@ -55,9 +55,7 @@ class SettingResource extends Resource
                     ->dateTime()
                     ->sortable(),
             ])
-            ->filters([
-                //
-            ])
+            ->filters([])
             ->actions([
                 Tables\Actions\EditAction::make(),
             ])
